+++ src/pages/group/create-group/components/create-group-form/index.ts
@@ -1,0 +1,1 @@
+export { CreateGroupForm } from './CreateGroupForm'
