+++ src/pages/group/create-group/components/index.ts
@@ -1,0 +1,1 @@
+export * from './create-group-form'
