+++ src/pages/group/create-group/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateGroupPage } from './CreatGroupPage'
