+++ src/pages/group/default/components/index.ts
@@ -1,4 +1,3 @@
 export * from './exit-group-button'
 export * from './group-profile'
 export * from './management'
-export * from './navigation'
