--- src/pages/group/default/components/navigation/index.ts
@@ -1,1 +1,0 @@
-export { GroupNavigation } from './Navigation'
