+++ src/pages/group/index.ts
@@ -1,1 +1,4 @@
 export * from './default'
+export * from './question-management'
+export * from './create-group'
+export * from './member-management'
