+++ src/pages/group/member-management/components/index.ts
@@ -1,0 +1,1 @@
+export * from './member-table'
