+++ src/pages/group/member-management/components/member-table/change-leader-button/index.ts
@@ -1,0 +1,1 @@
+export { ChangeLeaderButton } from './ChangeLeaderButton'
