+++ src/pages/group/member-management/components/member-table/expel-member-button/index.ts
@@ -1,0 +1,1 @@
+export { ExpelMemberButton } from './ExpelMemberButton'
