+++ src/pages/group/member-management/components/member-table/group-table/index.ts
@@ -1,0 +1,1 @@
+export { GroupTable } from './GroupTable'
