+++ src/pages/group/member-management/components/member-table/header/index.ts
@@ -1,0 +1,1 @@
+export { GroupMemberHeader } from './Header'
