+++ src/pages/group/member-management/components/member-table/index.ts
@@ -1,0 +1,1 @@
+export { GroupMemberTable } from './MemberTable'
