+++ src/pages/group/member-management/components/member-table/table-pagination/index.ts
@@ -1,0 +1,1 @@
+export { TablePagination } from './TablePagination'
