+++ src/pages/group/member-management/index.ts
@@ -1,0 +1,1 @@
+export { default as MemberManagementPage } from './MemberManagementPage'
