+++ src/pages/group/question-management/components/question-list/index.ts
@@ -1,0 +1,1 @@
+export { GroupQuestionList } from './QuestionList'
