+++ src/pages/group/question-management/index.ts
@@ -1,0 +1,1 @@
+export { default as QuestionManagementPage } from './QuestionManagementPage'
